added a control to check that Content key exists

+++ Core/Block/AlBlockManagerBootstrapThumbnailsBlock.php
@@ -40,22 +40,24 @@ class AlBlockManagerBootstrapThumbnailsBlock extends AlBlockManagerContainer
     
     protected function edit(array $values)
     {
-        $data = json_decode($values['Content'], true); 
-        $savedValues = AlBlockManagerJsonBlock::decodeJsonContent($this->alBlock);
-        
-        if ($data["operation"] == "add") {
-            $savedValues[] = $data["value"];
-            $values = array("Content" => json_encode($savedValues));
-        }
-        
-        if ($data["operation"] == "remove") {
-            unset($savedValues[$data["item"]]);
-            
-            $blocksRepository = $this->container->get('alpha_lemon_cms.factory_repository');
-            $repository = $blocksRepository->createRepository('Block');
-            $repository->deleteIncludedBlocks($data["slotName"]);
-            
-            $values = array("Content" => json_encode($savedValues));
+        if (array_key_exists('Content', $values)) {
+            $data = json_decode($values['Content'], true); 
+            $savedValues = AlBlockManagerJsonBlock::decodeJsonContent($this->alBlock);
+
+            if ($data["operation"] == "add") {
+                $savedValues[] = $data["value"];
+                $values = array("Content" => json_encode($savedValues));
+            }
+
+            if ($data["operation"] == "remove") {
+                unset($savedValues[$data["item"]]);
+
+                $blocksRepository = $this->container->get('alpha_lemon_cms.factory_repository');
+                $repository = $blocksRepository->createRepository('Block');
+                $repository->deleteIncludedBlocks($data["slotName"]);
+
+                $values = array("Content" => json_encode($savedValues));
+            }
         }
         
         return parent::edit($values);
